modulus: nest build artifacts in directories
It is more proper for tars to contain a directory which in turn contains
all of the build artifacts than for the tar to contain the files
directly, since this results in files populating the current working
directory during extraction. Compressing the current directory rather
than a subdirectory also results in racey behavior where `tar` tries to
compress the archive it is creating and fails.