modulus: revert using host's /proc/config.gz
This commit reverts a change that introduced a bug where the host's
/proc/config.gz was used. This would force the compilation to always
match the kernel version of the host. This is not what we want because
it prevents us from being able to compile for future versions, e.g. on
shutdown for the next OS version so that the modules are ready, or e.g.
compiling modules in CI running Ubuntu (as we do).

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>